課題3_コード修正
