+++ src/App.js
@@ -1,20 +1,16 @@
 import { Route, createBrowserRouter, createRoutesFromElements } from 'react-router-dom';
 
 import Header from './Header';
-// import ListPost from './ListPost';
-import ListPostAPI from './ListPostAPI';
+import ListPost from './ListPost';
 import Contact from './Contact';
-// import PostDitail from './PostDitail';
-import PostDitailAPI from './PostDitailAPI';
+import PostDitail from './PostDitail';
 
 const App = createBrowserRouter(
   createRoutesFromElements(
     <Route path="/" element={<Header/>}>
-      {/* <Route path="/" element={<ListPost />} /> */}
-      <Route path="/" element={<ListPostAPI />} />
+      <Route path="/" element={<ListPost />} />
       <Route path="contact" element={<Contact />} />
-      {/* <Route path="post/:postId" element={<PostDitail/>} /> */}
-      <Route path="post/:postId" element={<PostDitailAPI/>} />
+      <Route path="post/:postId" element={<PostDitail/>} />
     </Route>
   )
 );
